Dr Mario is alive
-Additions
1. Added DMA support without need of hard coded signals
2. Added Debug Enable for CPU (Disables Debug Printing)
3. Added frame timer, currently runs around 63fps instead of 59.975fps but its close enough for now
4. added background rendering and the entire FIFO rendering pipeline (only works 100% on DR Mario rn)

-Fixes/Quality Changes
1. General bug fixes around the emu

-Dev Notes
1. Generally just working on the PPU, chugging along as its alot of reading
2. Still need to remove SDL3, but im just working on the PPU

+++ src/inc/cpu.h
@@ -54,7 +54,7 @@ InstructionProcess instructionGetProcessor(instructionType type);
 uint16_t cpuReadReg(registerType rt);
 void cpuSetReg(registerType rt, uint16_t value);
 
-uint16_t cpuGetInterruptReg();
+uint8_t cpuGetInterruptReg();
 void cpuSetInterruptRegister(uint8_t n);
 
 uint8_t cpuReadReg8(registerType rt);
